split _other.py into several files
