+++ src/pyguarantees/_constraints/__init__.py
@@ -5,6 +5,11 @@ from ._collections import _IsSet, _IsFrozenSet, _IsDict, _IsList, _IsRange, \
 from ._string import _IsStr
 from ._binary import _IsMemoryView, _IsBytes, _IsByteArray
 from ._boolean import _IsBool
-from ._other import _NoOp, _IsClass, _IsNone, _IsUnion, _Cls, _Self
+from ._union import _IsUnion
+from ._class import _IsClass
+from ._none import  _IsNone
+from ._no_op import _NoOp
+from ._self import _Self
+from ._cls import _Cls
 from ._dynamic_check import _DynamicCheck
 import pyguarantees._constraints._util
