+++ src/pyguarantees/_constraints/_union.py
@@ -1,75 +1,14 @@
 from dataclasses import dataclass
-from typing import Type, List, Union
+from typing import List, Union
 
-from ._base import _Guarantee, _TypeGuarantee
+from ._base import _TypeGuarantee
+from ._no_op import _NoOp
 from pyguarantees._constraints._util.error_handeling import \
     handle_error
 from pyguarantees._constraints._util.typenames import \
     get_arg_type_name, get_type_name
 from pyguarantees._constraints._util.common_checks import \
     enforce_dynamic_checks, check_forbidden_values
-
-
-@dataclass
-class _NoOp(_Guarantee):
-    def __post_init__(self):
-        self.guarantee_name = "NoOp"
-        self.guaranteed_type = None
-
-    def enforce(self, arg):
-        return arg
-
-
-@dataclass
-class _Self(_NoOp):
-    def __post_init__(self):
-        self.guarantee_name = "Self"
-        self.guaranteed_type = None
-
-
-@dataclass
-class _Cls(_NoOp):
-    def __post_init__(self):
-        self.guarantee_name = "Cls"
-        self.guaranteed_type = None
-
-
-@dataclass
-class _IsClass(_TypeGuarantee):
-    class_type: Type = None
-
-    def __post_init__(self):
-        self.guarantee_name = "IsClass"
-        self.guaranteed_type = self.class_type
-
-    def enforce(self, arg):
-        arg = _check_isclass(arg, self)
-        arg = check_forbidden_values(arg, self)
-        enforce_dynamic_checks(arg, self)
-        return arg
-
-
-@dataclass
-class _IsNone(_TypeGuarantee):
-    def __post_init__(self):
-        self.guarantee_name = "IsNone"
-        self.guaranteed_type = None
-
-    def enforce(self, arg):
-        if arg is None:
-            enforce_dynamic_checks(arg, self)
-            return
-
-        handle_error(
-            where=self.where,
-            type_or_value="type",
-            guarantee=self,
-            parameter_name=self.parameter_name,
-            what_dict={
-                "should_type": "None",
-                "actual_type": get_arg_type_name(arg)
-            }
-        )
 
 
 @dataclass
@@ -160,19 +99,3 @@ class _IsUnion(_TypeGuarantee):
             parameter_name=self.parameter_name,
             what_dict={"error": "IsUnion may not contain IsUnion."}
         )
-
-
-def _check_isclass(arg: object, guarantee: _IsClass) -> object:
-    if guarantee.class_type is None or isinstance(arg, guarantee.class_type):
-        return arg
-
-    handle_error(
-        where=guarantee.where,
-        type_or_value="type",
-        guarantee=guarantee,
-        parameter_name=guarantee.parameter_name,
-        what_dict={
-            "should_type": str(guarantee.class_type),
-            "actual_type": get_arg_type_name(arg)
-        }
-    )
